% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinb.R
\name{fit.zinb}
\alias{fit.zinb}
\title{Zero-Inflated Negative Binomial Regression}
\usage{
fit.zinb(X, y, tol = 1e-04, maxIter = 1000, initial = NULL)
}
\arguments{
\item{X}{A \eqn{n} by \eqn{p} design matrix. Each row is a vector of observation with \eqn{p} covariates. Number of observations \eqn{n} must be greater than number of covariates \eqn{p}.}

\item{y}{An \eqn{n}-dimensional response vector.}

\item{tol}{(\strong{optional}) The tolerance level in the iterative estimation procedure. The defalut value is 1e-4.}

\item{maxIter}{(\strong{optional}) Maximum number of iterations. Default is 1000.}

\item{initial}{(\strong{optional}) A list specifies the initial point of the algorithm \code{list(beta=..., gamma=..., theta=...)}. Default is NULL.}
}
\value{
A list including the following terms will be returned:
\describe{
\item{\code{beta}}{The coefficients for the Negative Binomial Model.}
\item{\code{gamma}}{The coefficients for the Zero Inflated Model.}
\item{\code{theta}}{The reciprocal of the size parameter for the Negative Binomial Distribution.}
\item{\code{vcov}}{The covariance matrix for all estimated paramters.}
\item{\code{loglikelihood}}{The log-likelihood of the estimated model.}
}
}
\description{
Conduct zero-inflated negative binomial regression for data. Estimate paramters and their covariance using EM algorithm.
}
\examples{
n <- 100
X <- matrix(c(rep(1, 50), rbinom(50, 1, 0.5)), 50, 2)
y <- sample(c(rnbinom(20, 10, 0.5), rep(0, 30)))
fit.mod <- fit.zinb(X, y)
fit.mod$beta
}
