% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinb.R
\name{LRTnD}
\alias{LRTnD}
\title{Likelihood Ratio Test for Zero-Inflated Negative Binomial Regression (Multiple Responses)}
\usage{
LRTnD(X, Y, inflation = NULL, add_intercept = TRUE, ...)
}
\arguments{
\item{X}{An \eqn{n} by \eqn{p} design matrix. Each row is a vector of observation with \eqn{p} covariates. Number of observations \eqn{n} must be greater than number of covariates \eqn{p}.}

\item{Y}{An \eqn{n} by \eqn{q} response matrix.}

\item{inflation}{(\strong{optional}) A \eqn{q}-dimensional logical vector indicating whether the inflated model should be used for each test. If set to NULL, use the inflated model for all tests. Default is NULL.}

\item{add_intercept}{(\strong{optional}) Whether an intercept term should be added to the design matrix X. If so, the first element of \code{beta} and \code{gamma} would be the intercept. Default is TRUE.}

\item{...}{(\strong{optional}) Other optional parameters for \code{fit.zinb}.}
}
\value{
A list including the following terms will be returned:
\describe{
\item{\code{Chisq}}{A vector of calculated Chisq statistics.}
\item{\code{df}}{The degree of freedom of the tests.}
\item{\code{p.value}}{A vector of computed p-values.}
\item{\code{q.value}}{A vector of adjusted p-values using the Benjamini-Hochberg (BH) correction.}
}
}
\description{
Testing the intercept model against the model fitted by the design matrix. It is recommended to include the intercept in X or set \code{add_intercept} as TRUE.
}
\examples{
n <- 100
X <- matrix(c(rep(1, 50), rbinom(50, 1, 0.5)), 50, 2)
y <- sample(c(rnbinom(20, 10, 0.5), rep(0, 30)))
Y <- matrix(c(y, sample(y)), 50, 2)
result <- LRTnD(X, Y, add_intercept=FALSE)
result$p.value
}
