% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinb.R
\name{LRTnD}
\alias{LRTnD}
\title{Likelihood Ratio Test for Zero-Inflated Negative Binomial Regression (Multiple Responses)}
\usage{
LRTnD(X, y, ...)
}
\arguments{
\item{X}{A \eqn{n} by \eqn{p} design matrix. Each row is a vector of observation with \eqn{p} covariates. Number of observations \eqn{n} must be greater than number of covariates \eqn{p}.}

\item{...}{(\strong{optional}) Other optional parameters for \code{fit.zinb}.}

\item{Y}{An \eqn{n} by \eqn{p} response matrix.}
}
\value{
A list including the following terms will be returned:
\describe{
\item{\code{Chisq}}{A vector of calculated Chisq statistics.}
\item{\code{df}}{The degree of freedom of this test.}
\item{\code{p.value}}{A vector of computed p-values.}
}
}
\description{
Testing the intercept model against the model fitted by the design matrix. It is recommended to include the intercept in X.
}
\examples{
n <- 100
X <- matrix(c(rep(1, 50), rbinom(50, 1, 0.5)), 50, 2)
y <- sample(c(rnbinom(20, 10, 0.5), rep(0, 30)))
Y <- matrix(c(y, sample(y)), 50, 2)
result <- LRTnD(X, Y)
result$p.value
}
